+++ src/real.cpp
@@ -1501,6 +1501,19 @@ real &real::trunc()
     return self_mpfr_unary_nornd(::mpfr_trunc);
 }
 
+// Set to n*2**e.
+real &set_ui_2exp(real &r, unsigned long n, ::mpfr_exp_t e)
+{
+    ::mpfr_set_ui_2exp(r._get_mpfr_t(), n, e, MPFR_RNDN);
+    return r;
+}
+
+real &set_si_2exp(real &r, long n, ::mpfr_exp_t e)
+{
+    ::mpfr_set_si_2exp(r._get_mpfr_t(), n, e, MPFR_RNDN);
+    return r;
+}
+
 // Implementation bits for in-place addition.
 namespace detail
 {