+++ doc/changelog.rst
@@ -7,6 +7,9 @@ Changelog
 New
 ~~~
 
+- Add the :cpp:func:`~mppp::set_ui_2exp()` and :cpp:func:`~mppp::set_si_2exp()`
+  functions for :cpp:class:`~mppp::real`
+  (`#242 <https://github.com/bluescarni/mppp/pull/242>`__).
 - The mp++ library can now be built with link-time
   optimisations enabled
   (`#240 <https://github.com/bluescarni/mppp/pull/240>`__).